Update speed-Patrick-Jakubowski.asm
Now runs in 54, 102 under par.

Two tweaks, It now enters the two ones before inboxing, this means we don't have to COPYFROM.  Should have seen this one from the start, oops.  Second, because of this tweak occuring in 4 places (initial, then second run in each possible outcome of the first), and it'd fail on the second pass through (which I wasn't sure of which it was because I linked back to a first run check of 13), I now have 3 separate methods for second pass checks of 13 8 and 5, which inboxes first and doesn't bother trying to put more 1's on the stack (because there are only ever 2 numbers to check against).  So now the first tweak can be applied universally, and we get a grand finale cycles of 54! 102 under par.