Add new solution (technically two) for 40-Prime-Factory.
This caches a table of the existing primes for fast checking, in a way that theoretically works for arbitrary input but effectively won't have to generate a table at all for the test data. I've included both the version where the table is pre-filled, and one where it's completely generated.